tests/submission/Infraware/Forms/contents/Forms/output_checkvalidity.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>output_checkValidity</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="" 
		  id="input_form">
	 <output name=o id='output_id'></output>
	</form>
	<script>

		var output = document.getElementById("output_id");

		try
		{
			var ret = output.checkValidity();

			test(function() {
				assert_equals(ret, true, "calling of checkValidity method is failed.");
			});
		}
		catch (e) {
			test(function() {
				assert_unreached("Error is raised.");
			});
		}

	</script>

 </body>
</html>
